+++ src/scheduler/event.py
@@ -1,28 +1,25 @@
+# For use with type hints,,,
+from __future__ import annotations
+
 import datetime
+from scheduler.defaults import EVENT_DURATION
 
 
 class Event:
     """
     A unique campus event.
-
-    Authors
-    -------
-    Augustine Joy
     """
 
     def __init__(
         self,
-        id: int,
         title: str,
         date: str,
         time: str,
         location: str,
     ):
         """
-        Attributes
+        Parameters
         ----------
-        id: int
-            A unique event ID
         title: str
             The title of the event
         date: str
@@ -34,85 +31,100 @@ class Event:
         location: str
             The location of the event
         """
-
-        self.id = id
         self.title = title
         self.location = location
         self.date = date
         self.time = time
 
-        # Checks whether date and time are valid inputs and returns None if its invalid.
+        # Checks whether date and time are valid inputs
         try:
             self.start_time = datetime.datetime.strptime(
                 f"{date} {time}", "%Y-%m-%d %H:%M"
             )
-            self.end_time = self.start_time + datetime.timedelta(hours=1)
+            # Create an event end time
+            # TODO: Make duration a customizable parameter
+            self.end_time = self.start_time + datetime.timedelta(seconds=EVENT_DURATION)
         except ValueError:
             raise ValueError(f"Invalid date '{date}' or time '{time}'")
 
-    # Function to check whether new event collides with other existing events.
-    def collides_with(self, other):
+        @property
+        def id(self):
+            """
+            A unique event ID.
+            """
+            return self._id
+
+        @id.setter
+        def id(self, value: int):
+            """
+            Set the unique event ID value.
+
+            Parameters
+            ----------
+            value: int
+                Value to set as the event ID
+            """
+            if isinstance(value, bool) or not isinstance(value, int):
+                raise TypeError(f"Cannot set ID to type {type(value)}")
+            self._id = value
+
+    def collides_with(self, other: Event):
+        """
+        Checks if two events overlap in time.
+
+        Parameters
+        ----------
+        other: Event
+            Another Event object
+        """
         return self.start_time < other.end_time and other.start_time < self.end_time
 
-    def __eq__(self, other):
+    def __eq__(self, other: Event):
+        """
+        Determines if two instances of Event are equivalent. Equality is defined as exactly matching attributes.
+        """
         if not isinstance(other, Event):
-            return NotImplemented
-        return self.start_time == other.start_time
-
-    def __ne__(self, other):
-        if not isinstance(other, Event):
-            return NotImplemented
-        return self.start_time != other.start_time
-
-    def __gt__(self, other):
-        if not isinstance(other, Event):
-            return NotImplemented
-        return self.start_time > other.start_time
-
-    def __lt__(self, other):
-        if not isinstance(other, Event):
-            return NotImplemented
-        return self.start_time < other.start_time
+            raise TypeError(
+                f"Cannot establish equality between Event and {type(other)}"
+            )
+        checked_attributes = ["title", "date", "time", "location", "id"]
+        filter_attributes = lambda obj: {
+            k: v for k, v in vars(obj).items() if k in checked_attributes
+        }
+        return filter_attributes(self) == filter_attributes(other)
 
 
 class EventNode(Event):
     """
     A unique campus event node for storage within a linked list.
-
-    Attributes
-    ----------
-    id: int
-        A unique event ID
-    title: str
-        The title of the event
-    date: str
-        The date the event will take place in the form YYYY-MM-DD
-    time: str
-        The time the event will take place in the form HH:MM
-    datetime: tuple | None
-        A tuple composed of the inputted date and time
-    location: str
-        The location of the event
-    next: None | Event
-        The next event in a linked list of events
-        If None, indicates the end of the list
-
-    Authors
-    -------
-    Stephanie Bie
     """
 
     def __init__(
         self,
-        id: int,
         title: str,
         date: str,
         time: str,
         location: str,
-        next: None | "EventNode" = None,
+        next: EventNode | None = None,
     ):
+        """
+        Parameters
+        ----------
+        title: str
+            The title of the event
+        date: str
+            The date the event will take place in the form YYYY-MM-DD
+        time: str
+            The time the event will take place in the form HH:MM
+        datetime: tuple | None
+            A tuple composed of the inputted date and time
+        location: str
+            The location of the event
+        next: None | EventNode
+            The next event in a linked list of events
+            If None, indicates the end of the list
+        """
         super().__init__(
-            id=id,
             title=title,
             date=date,
             time=time,